src/app/search-product/search-product.component.ts
import { Component, OnInit } from '@angular/core';
import { ProductModel } from '../models/product-model';
import { ProductService } from '../services/product.service';
import { ActivatedRoute } from '@angular/router';
import { CartService } from '../services/cart.service';

@Component({
  selector: 'app-search-product',
  templateUrl: './search-product.component.html',
  styleUrls: ['./search-product.component.css']
})
export class SearchProductComponent implements OnInit {

  name;
  products: ProductModel;

  constructor(private productService: ProductService, private activatedRoute: ActivatedRoute, private cartService: CartService) { }

  ngOnInit() {
    this.name = this.activatedRoute.snapshot.paramMap.get('name');
    this.productService.searchProduct(this.name).subscribe(res => this.products = res);

  }

  addToCart(product) {
    window.alert("Product has been added to the cart !");
    this.cartService.addToCart(product);
  }

}
